+++ lib/Time-1.6.1/examples/TimeNTP/TimeNTP.ino
@@ -91,7 +91,7 @@ time_t getNtpTime()
   while (Udp.parsePacket() > 0) ; // discard any previously received packets
   Serial.println("Transmit NTP Request");
   sendNTPpacket(timeServer);
-  uint32_t beginWait = millis();
+  unsigned long beginWait = millis();
   while (millis() - beginWait < 1500) {
     int size = Udp.parsePacket();
     if (size >= NTP_PACKET_SIZE) {
