+++ src/main.h
@@ -17,7 +17,7 @@ extern "C" {
 }
 
 uint8_t init_OTA();
-uint32_t getTime();
+unsigned long getTime();
 void setupSPIslave();
 char* timeStamp();
 char* dateStamp();
@@ -32,7 +32,7 @@ void handleNotFound();
 void ISRwatchDog();
 void waitForData();
 void minProc();
-void watchWait(uint32_t);
+void watchWait(unsigned long);
 void joinNet();
 void setupTime();
 void updateEnergyFile();
@@ -88,10 +88,10 @@ uint8_t scanSec, badSumCount = 0;
 uint16_t i, oldYear, htmlLen, CstringPtr, syncDelay = 800;
 uint16_t localPort = 4210;          //  must match port assigned in "pulse"
 uint16_t loopCount = 0, missedCycle = 0;
-uint32_t t0, t1, t2, t3, startMillis, startSeconds, lastScan, startOutage;
-uint32_t t_lastData, t_scan, t_scan_max, waiting, WWmin, WWmax, WDmin, WDmax;
-uint32_t loopStart, loopTime, LTmin, LTmax, progLoopTime;
-uint32_t wfdStart, wfdTime, wfdPrev, WFDmin, WFDmax;
+unsigned long t0, t1, t2, t3, startMillis, startSeconds, lastScan, startOutage;
+unsigned long t_lastData, t_scan, t_scan_max, waiting, WWmin, WWmax, WDmin, WDmax;
+unsigned long loopStart, loopTime, LTmin, LTmax, progLoopTime;
+unsigned long wfdStart, wfdTime, wfdPrev, WFDmin, WFDmax;
 
 float Wrms[MAX_CCTS+1];					// Sum of sampled V*I
 /*  RMS1:
